pipelines/19-mage-ai/mage_project/requirements.txt
kafka-python-ng>=2.0.2
duckdb>=0.10.0
pyarrow>=14.0.0
